Remove shiny_app
At the start of the project my intention was to create a shiny_app out of the plots and with that in minds put a highly preliminary skeleton for experimenting.

I had been intending to make use of the new features in the bslib package https://rstudio.github.io/bslib/articles/dashboards/index.html

However when I ran into difficulty installing and with the help of and instructor it was found that the features I wanted to use were part of a *development* version 0.5.1.9000 not the current stable release 0.5.1. I did not want to proceed for reproducibility and unsure whether there would turn out to be other issues related to the unstable code.

I left it for few days to see if there was any project activity but there was none to suggest a stable release would come during the window of development time for this project.

Other delays elsewhere meant I abandoned the app idea but I did pivot and instead created an interactive shiny document for a cancer_risk_tool.